requirements.txt
streamlit
feedparser
requests
google-generativeai